Move the organization resources to their own file
