+++ h/traversal/roots.py
@@ -70,7 +70,7 @@ from h.auth import role
 from h.auth.util import client_authority
 from h.exceptions import InvalidUserId
 from h.interfaces import IGroupService
-from h.models import AuthClient, Organization
+from h.models import AuthClient
 from h.traversal import contexts
 
 
@@ -147,50 +147,6 @@ class BulkAPIRoot(RootFactory):
 
     # Currently only LMS uses this end-point
     __acl__ = [(Allow, "client_authority:lms.hypothes.is", "bulk_action")]
-
-
-class OrganizationRoot(RootFactory):
-    """
-    Root factory for routes whose context is an :py:class:`h.traversal.OrganizationContext`.
-
-    FIXME: This class should return OrganizationContext objects, not Organization
-    objects.
-
-    """
-
-    def __getitem__(self, pubid):
-        try:
-            org = self.request.db.query(Organization).filter_by(pubid=pubid).one()
-
-            # Inherit global ACL. See comments in :py:class`h.traversal.AuthClientRoot`.
-            org.__parent__ = Root(self.request)
-
-            return org
-        except sqlalchemy.orm.exc.NoResultFound:
-            raise KeyError()
-
-
-class OrganizationLogoRoot(RootFactory):
-    """
-    Root factory for routes whose context is an :py:class:`h.traversal.OrganizationLogoContext`.
-
-    FIXME: This class should return OrganizationLogoContext objects, not
-    organization logos.
-
-    """
-
-    def __init__(self, request):
-        super().__init__(request)
-        self.organization_factory = OrganizationRoot(self.request)
-
-    def __getitem__(self, pubid):
-        # This will raise KeyError if the organization doesn't exist.
-        organization = self.organization_factory[pubid]
-
-        if not organization.logo:
-            raise KeyError()
-
-        return organization.logo
 
 
 class GroupRoot(RootFactory):
