+++ tests/h/traversal/contexts_test.py
@@ -10,7 +10,6 @@ from h.traversal.contexts import (
     AnnotationContext,
     GroupContext,
     GroupUpsertContext,
-    OrganizationContext,
     UserContext,
 )
 
@@ -292,58 +291,6 @@ class TestGroupContext:
         assert group_context.organization is None
 
 
-@pytest.mark.usefixtures("organization_routes")
-class TestOrganizationContext:
-    def test_it_returns_organization_model_as_property(
-        self, factories, pyramid_request
-    ):
-        organization = factories.Organization()
-
-        organization_context = OrganizationContext(organization, pyramid_request)
-
-        assert organization_context.organization == organization
-
-    def test_it_returns_pubid_as_id(self, factories, pyramid_request):
-        organization = factories.Organization()
-
-        organization_context = OrganizationContext(organization, pyramid_request)
-
-        assert organization_context.id != organization.id
-        assert organization_context.id == organization.pubid
-
-    def test_it_returns_links_property(self, factories, pyramid_request):
-        organization = factories.Organization()
-
-        organization_context = OrganizationContext(organization, pyramid_request)
-
-        assert organization_context.links == {}
-
-    def test_it_returns_logo_property_as_route_url(self, factories, pyramid_request):
-        fake_logo = "<svg>H</svg>"
-        pyramid_request.route_url = mock.Mock()
-
-        organization = factories.Organization(logo=fake_logo)
-
-        organization_context = OrganizationContext(organization, pyramid_request)
-        logo = organization_context.logo
-
-        pyramid_request.route_url.assert_called_with(
-            "organization_logo", pubid=organization.pubid
-        )
-        assert logo is not None
-
-    def test_it_returns_none_for_logo_if_no_logo(self, factories, pyramid_request):
-        pyramid_request.route_url = mock.Mock()
-
-        organization = factories.Organization(logo=None)
-
-        organization_context = OrganizationContext(organization, pyramid_request)
-        logo = organization_context.logo
-
-        pyramid_request.route_url.assert_not_called
-        assert logo is None
-
-
 @pytest.mark.usefixtures("links_svc")
 class TestGroupUpsertContext:
     def test_acl_applies_root_upsert_to_user_role_when_no_group(
@@ -454,8 +401,3 @@ def links_svc(pyramid_config):
     svc = mock.create_autospec(GroupLinksService, spec_set=True, instance=True)
     pyramid_config.register_service(svc, name="group_links")
     return svc
-
-
-@pytest.fixture
-def organization_routes(pyramid_config):
-    pyramid_config.add_route("organization_logo", "/organization/{pubid}/logo")
